Fix clang compilation warnings
- listjob: drop the unused Q_D in setDetails
- kfileplacesview: remove a no-op QRect::adjust whose sub-pixel args truncate to 0
- kfilewidget: cast the std::array size to int in the slider bounds assert to avoid a sign-compare
- kfileitemdelegate: drop unused parameter names
- renamedialog, threadtest, threadconnectionbackendtest: drop lambda captures of const values usable without capture
- widgetsopenorexecutefilehandler: drop unused lambda captures
- testtrash: mark dirListerContainsDisplayName maybe_unused, it is only used in the no-QtDBus build

+++ src/filewidgets/kfilewidget.cpp
@@ -2276,7 +2276,7 @@ void KFileWidgetPrivate::slotDirOpIconSizeChanged(int size)
 {
     int sliderValue = m_iconSizeSlider->value();
     short currentIconSliderSize = m_stdIconSizes[sliderValue];
-    Q_ASSERT(sliderValue < m_stdIconSizes.size());
+    Q_ASSERT(sliderValue < static_cast<int>(m_stdIconSizes.size()));
     if (size == currentIconSliderSize)
         return;
     if (std::abs(size - currentIconSliderSize) != 1) {
